requirements.txt
# Core Cryptographic Libraries
pycryptodome==3.19.0        # AES-256 encryption
charm-crypto==0.43          # CP-ABE (Attribute-Based Encryption)
# Note: charm-crypto 0.43 is the latest available on Windows via pip

# Web Framework
fastapi==0.104.1            # REST API framework
uvicorn==0.24.0             # ASGI server
pydantic==2.5.0             # Data validation

# Database
sqlalchemy==2.0.23          # ORM
alembic==1.12.1             # Database migrations
psycopg2-binary==2.9.9      # PostgreSQL adapter (optional)

# Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4      # Password hashing
python-multipart==0.0.6     # File uploads

# Utilities
pyyaml==6.0.1               # Configuration files
python-dotenv==1.0.0        # Environment variables
click==8.1.7                # CLI interface
colorama==0.4.6             # Colored terminal output

# Testing
pytest==7.4.3               # Testing framework
pytest-cov==4.1.0           # Coverage reporting
pytest-asyncio==0.21.1      # Async testing
faker==20.1.0               # Fake data generation

# Development
black==23.11.0              # Code formatter
flake8==6.1.0               # Linting
mypy==1.7.1                 # Type checking
pre-commit==3.5.0           # Git hooks
